% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tertile_turtle.R
\name{tertile_turtle}
\alias{tertile_turtle}
\title{Tertile-turtle function}
\usage{
tertile_turtle(x = NULL, tertile = F)
}
\arguments{
\item{x}{a numeric vector (optional).}
}
\value{
An ASCII turtle. Optionally, returns an ordered factor corresponding to the tertile split of a given vector.
}
\description{
Returns a turtle (and, optionally, a tertile)
}
\examples{
\dontrun{
turtle();
turtle(1:10, tertile = T);
}
}
